IlyonAi-Wallet-assistant-main/server/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# LLM / Agent framework
langchain==0.3.1
langchain-openai==0.2.1
langchain-groq==0.2.1

# Blockchain
web3==7.3.0

# HTTP clients
requests==2.32.3
httpx==0.27.2

# Environment
python-dotenv==1.0.1

# Auth / database
sqlalchemy==2.0.36
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
email-validator==2.2.0
PyNaCl==1.5.0
base58==2.1.1
